Add Metavar[A] type class to provide default metavar names.
This change removes the "defaultMetvar" field from Argument[A], and
adds a new Metavar[A] type class to provide the same functionality.
The API impact of this is minimal: there is an extra type class
instance required when constructing opts, but a low-priority default
Metavar implementation ensures that this is unlikely to be a huge
burden.

The current Metavar implementations were created to match the existing
behavior (e.g. using "integer" for all integer types). I added a bunch
of extra implementations that users might be interested in, but we can
remove those if it seems too noisy.