Rewrite the normalization / arg parsing
In the previous version, we implemented the right-distribution property for
applicatives by normalizing up-front. This worked, but could cause exponential
blowup in some really-not-that-rare cases.

This change doesn't do any normalization up front, dealing with the issue at
parse time. There's some danger that we'd still expand the full tree after the
first argument is parsed: so we make an effort to keep the OrElse nodes as close
to the leaves as possible.